+++ lib/helpers.h
@@ -216,7 +216,7 @@ void thread_waker_del(struct thread_waker *);
 void threads_join_all(bool cancel);
 void thread_create_detach_prio(void (*)(void *), void *, const char *, int, const char *);
 void thread_create_looper(enum thread_looper_action (*f)(void), const char *scheduler, int priority,
-		const char *name, long long);
+		const char *name, int64_t);
 INLINE void thread_create_detach(void (*f)(void *), void *a, const char *name) {
 	thread_create_detach_prio(f, a, NULL, 0, name);
 }
